Refactor vocdoniSdk provider
